Fix Python 3 compatibility: Convert division results to int for range() functions
- Fixed range() TypeError in level.py, justice.py, taut.py, run_ock_new.py
- Disabled 60-minute time limit for indefinite leveling
- Updated Kaeval recall navigation to handle both 'The Baptismal Font' and 'The Stables'
- Fixed solo startup IndexError in log_cleric() function

+++ justice.py
@@ -90,7 +90,7 @@ class dhaven:
                 
                 if getpotion:
                     if potreq > 20:
-                        for i in range(max([potreq/20-1,0])):
+                        for i in range(max([int(potreq/20)-1,0])):
                             self.rod.write("buy %s %s\nempty bag basket\ndrop bag\n"%(20, self.potname.split()[2]))
                     else:
                         self.rod.write("buy %s %s\nempty bag basket\ndrop bag\n"%(20, self.potname.split()[2]))
@@ -138,7 +138,7 @@ class dhaven:
         print self.HP, self.MAXHP
 
         healN = 1200-int(self.HP)
-        for i in range(healN/600+1):
+        for i in range(int(healN/600)+1):
             self.waitcmd("say heal")
         
         #while True:
